refactor: CouponFacade 제거
- CouponFacade에서 참조하는 다른 Service가 제거 되어 Facade 계층이 불필요함
