+++ src/main/java/kr/hhplus/be/server/coupon/controller/CouponController.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.coupon.controller;
 
+import kr.hhplus.be.server.coupon.application.CouponService;
 import kr.hhplus.be.server.coupon.domain.Coupon;
-import kr.hhplus.be.server.coupon.facade.CouponFacade;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
@@ -12,11 +12,11 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CouponController implements CouponApi {
 
-    private final CouponFacade couponFacade;
+    private final CouponService couponService;
 
     @Override
     public ResponseEntity<CouponResponseDto> claimCoupon(Long userId, Long policyId) {
-        Coupon coupon = couponFacade.issueCoupon(userId, policyId);
+        Coupon coupon = couponService.issueCoupon(userId, policyId);
 
         CouponResponseDto dto = new CouponResponseDto(coupon.getId(), coupon.getDiscountAmount());
         return ResponseEntity.status(201).body(dto);
@@ -24,7 +24,7 @@ public class CouponController implements CouponApi {
 
     @Override
     public ResponseEntity<List<CouponResponseDto>> getUserCoupons(Long userId) {
-        List<Coupon> coupons = couponFacade.getCoupons(userId);
+        List<Coupon> coupons = couponService.getCoupons(userId);
         List<CouponResponseDto> result = coupons.stream()
                 .map(c -> new CouponResponseDto(c.getId(), c.getDiscountAmount()))
                 .toList();
