--- src/main/java/kr/hhplus/be/server/coupon/facade/CouponFacade.java
@@ -1,36 +1,0 @@
-package kr.hhplus.be.server.coupon.facade;
-
-import kr.hhplus.be.server.coupon.application.CouponService;
-import kr.hhplus.be.server.coupon.domain.Coupon;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Component
-@RequiredArgsConstructor
-public class CouponFacade {
-
-    private final CouponService couponService;
-
-    @Transactional
-    public Boolean tryIssue(Long userId, Long policyId) {
-        return couponService.tryIssue(userId, policyId);
-    }
-
-    @Transactional
-    public Coupon issueCoupon(Long userId, Long policyId) {
-        return couponService.issueCoupon(userId, policyId);
-    }
-
-    @Transactional
-    public Coupon useCoupon(Long userId, Long couponId) {
-        return couponService.useCoupon(couponId, userId);
-    }
-
-    @Transactional(readOnly = true)
-    public List<Coupon> getCoupons(Long userId) {
-        return couponService.getCoupons(userId);
-    }
-}
